+++ src/components/index.ts
@@ -1,0 +1,9 @@
+// Main components
+export { default as Map } from './Map';
+export { MapControls } from './MapControls';
+export { MapStyleSelector } from './MapStyleSelector';
+export { ColorSchemeSelector } from './ColorSchemeSelector';
+export { RecenterButton, RecenterButtonUI, triggerRecenter } from './RecenterButton';
+export { CountryLayer } from './CountryLayer';
+export { CityMarkers } from './CityMarkers';
+export { SimpleGallery } from './SimpleGallery';
