+++ src/hooks/useRecenter.ts
@@ -1,0 +1,10 @@
+import { useCallback } from 'react';
+import { triggerRecenter } from '../components';
+
+export function useRecenter() {
+  const handleRecenter = useCallback(() => {
+    triggerRecenter();
+  }, []);
+
+  return { handleRecenter };
+}
