+++ src/pages/index.tsx
@@ -1,16 +1,6 @@
 import dynamic from 'next/dynamic';
 import visitedData from '../../data/visited.json';
-
-interface City {
-  city: string;
-  country: string;
-  coordinates: [number, number];
-  photos?: string[];
-}
-
-interface VisitedData {
-  visited: City[];
-}
+import { VisitedData } from '../types';
 
 const Map = dynamic(() => import('../components/Map'), { ssr: false });
